Migrate firmware examples to pru-gcc
The firmware examples have been moved to gnupru
(https://github.com/dinuxbg/gnupru). The modifications done are:
1. header files - The register declarations for CT_CFG, CT_MBX and
CT_INTC have been changed.
2. fimware : main - __R30 and __R31 are not available. Instead
write_r30(), read_r30(), write_r31() and read_r31() have been used.
3. firmware: resource tables - modified according to the pru-gcc way of
generating elf files.
4. Makefiles - Modified to use pru-gcc

+++ examples/include/pru_virtio_ring.h
@@ -135,7 +135,7 @@ static inline void vring_init(struct vring *vr, uint32_t num, void *p,
 	vr->num = num;
 	vr->desc = p;
 	vr->avail = (void *)((char *)p + num*sizeof(struct vring_desc));
-	vr->used = (void *)(uintptr_t)(((uint64_t)&vr->avail->ring[num]
+	vr->used = (void *)(uintptr_t)(((uintptr_t)&vr->avail->ring[num]
 	                                + sizeof(uint16_t) + align-1) & ~(align - 1));
 }
